requirements.txt
asgiref==3.3.1
Django==3.1.5
Pillow==8.1.0
python-dotenv==0.15.0
pytz==2020.5
sqlparse==0.4.1
